fix warn

+++ services/distributeddataservice/service/data_share/common/rdb_delegate.cpp
@@ -35,7 +35,7 @@ enum REMIND_TIMER_ARGS : int32_t {
 };
 std::string RemindTimerFunc(const std::vector<std::string> &args)
 {
-    int size = args.size();
+    size_t size = args.size();
     if (size != ARGS_SIZE) {
         ZLOGE("RemindTimerFunc args size error, %{public}d", size);
         return "";
